FIX: Crash on startup due to ratelimit
If we start up multiple instances of the bot at the same time
the discord api will rate limit us and we will have to wait for the
specified amount of milliseconds.

As the expected keys are not present in the payload, json["url"]
will return nil and hence nil <> "?v=6&encoding=json" will crash.